+++ cmd/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
 ## [0.3.3]
+* Add aws log ingest count query config
 
 ## [0.3.2]
 * Bugfix: AzureResourceProperties.Other field saving to MongoDB as binary data